moved media's logic to a separate class
